pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Leo Graciano <leonardoferreiragraciano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.115.0"}
sqlalchemy = "^2.0.35"
pydantic-settings = "^2.5.2"
alembic = "^1.13.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 88
indent-width = 4
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]

[tool.ruff.format]
preview = true
docstring-code-format = true
quote-style = "double"

[tool.taskipy.tasks]
run = 'fastapi dev fast_zero/app.py'

lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"